+++ src/traits.rs
@@ -510,7 +510,6 @@ where
     /// [`wrapping_from_fixed`]: trait.FromFixed.html#tymethod.wrapping_from_fixed
     fn wrapping_to_num<Dst: FromFixed>(self) -> Dst;
 
-    #[cfg(feature = "unwrapped")]
     /// Creates a fixed-point number from another number, panicking on overflow.
     ///
     /// # Panics
@@ -518,7 +517,6 @@ where
     /// Panics if the value does not fit.
     fn unwrapped_from_num<Src: ToFixed>(src: Src) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Converts a fixed-point number to another number, panicking on overflow.
     ///
     /// # Panics
@@ -732,60 +730,36 @@ where
     /// with ties rounded to even, and wrapping on overflow.
     fn wrapping_round_ties_to_even(self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped ceil. Rounds to the next integer towards +∞,
     /// panicking on overflow.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_ceil(self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped floor. Rounds to the next integer towards −∞,
     /// panicking on overflow.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_floor(self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped round. Rounds to the next integer to the nearest,
     /// with ties rounded away from zero, and panicking on overflow.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_round(self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped round. Rounds to the next integer to the nearest,
     /// with ties rounded to even, and panicking on overflow.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_round_ties_to_even(self) -> Self;
 
     /// Overflowing ceil. Rounds to the next integer towards +∞.
@@ -1140,192 +1114,108 @@ where
     /// bits, then shifts and returns the number.
     fn wrapping_shr(self, rhs: u32) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped negation. Returns the negated value, panicking on overflow.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_neg(self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped addition. Returns the sum, panicking on overflow.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_add(self, rhs: Self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped subtraction. Returns the difference, panicking on overflow.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_sub(self, rhs: Self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped multiplication. Returns the product, panicking on overflow.
-    ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
     ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_mul(self, rhs: Self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped division. Returns the quotient, panicking on overflow.
-    ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
     ///
     /// # Panics
     ///
     /// Panics if the divisor is zero or if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_div(self, rhs: Self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped reciprocal. Returns reciprocal, panicking on overflow.
-    ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
     ///
     /// # Panics
     ///
     /// Panics if `self` is zero or on overflow.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_recip(self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped multiply and add. Returns `self` × `mul` + `add`, panicking on overflow.
-    ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
     ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_mul_add(self, mul: Self, add: Self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped Euclidean division. Returns the quotient, panicking on overflow.
-    ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
     ///
     /// # Panics
     ///
     /// Panics if the divisor is zero or if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_div_euclid(self, rhs: Self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped multiplication by an integer. Returns the product, panicking on overflow.
-    ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
     ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_mul_int(self, rhs: Self::Bits) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped division by an integer. Returns the quotient, panicking on overflow.
     ///
     /// Overflow can only occur when dividing the minimum value by −1.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the divisor is zero or if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_div_int(self, rhs: Self::Bits) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped Euclidean division by an integer. Returns the
     /// quotient, panicking on overflow.
     ///
     /// Overflow can only occur when dividing the minimum value by −1.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the divisor is zero or if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_div_euclid_int(self, rhs: Self::Bits) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped remainder for Euclidean division by an integer.
     /// Returns the remainder, panicking on overflow.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the divisor is zero or if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_rem_euclid_int(self, rhs: Self::Bits) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped shift left. Panics if `rhs` ≥ the number of bits.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if `rhs` ≥ the number of bits.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_shl(self, rhs: u32) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped shift right. Panics if `rhs` ≥ the number of bits.
-    ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
     ///
     /// # Panics
     ///
     /// Panics if `rhs` ≥ the number of bits.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_shr(self, rhs: u32) -> Self;
 
     /// Overflowing negation.
@@ -1588,22 +1478,15 @@ pub trait FixedSigned: Fixed + Neg<Output = Self> {
     ///         integer bits, such that it cannot hold the value −1.
     fn wrapping_signum(self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped absolute value. Returns the absolute value, panicking on overflow.
     ///
     /// Overflow can only occur when trying to find the absolute value of the minimum value.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_abs(self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Unwrapped signum. Returns a number representing the sign of
     /// `self`, panicking on overflow.
     ///
@@ -1613,14 +1496,9 @@ pub trait FixedSigned: Fixed + Neg<Output = Self> {
     ///   * if the value is negative and the fixed-point number has zero
     ///         integer bits, such that it cannot hold the value −1.
     ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
-    ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_signum(self) -> Self;
 
     /// Overflowing absolute value.
@@ -1685,18 +1563,12 @@ pub trait FixedUnsigned: Fixed {
     /// to 0 if the next power of two is too large to represent.
     fn wrapping_next_power_of_two(self) -> Self;
 
-    #[cfg(feature = "unwrapped")]
     /// Returns the smallest power of two that is ≥ `self`, panicking
     /// if the next power of two is too large to represent.
-    ///
-    /// This method is only available when the [`unwrapped`
-    /// experimental feature][exp] is enabled.
     ///
     /// # Panics
     ///
     /// Panics if the result does not fit.
-    ///
-    /// [exp]: ../index.html#experimental-optional-features
     fn unwrapped_next_power_of_two(self) -> Self;
 }
 
@@ -2413,9 +2285,7 @@ macro_rules! impl_fixed {
             trait_delegate! { fn saturating_to_num<Dst: FromFixed>(self) -> Dst }
             trait_delegate! { fn wrapping_from_num<Src: ToFixed>(val: Src) -> Self }
             trait_delegate! { fn wrapping_to_num<Dst: FromFixed>(self) -> Dst }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_from_num<Src: ToFixed>(val: Src) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_to_num<Dst: FromFixed>(self) -> Dst }
             trait_delegate! { fn overflowing_from_num<Src: ToFixed>(val: Src) -> (Self, bool) }
             trait_delegate! { fn overflowing_to_num<Dst: FromFixed>(self) -> (Dst, bool) }
@@ -2477,13 +2347,9 @@ macro_rules! impl_fixed {
             trait_delegate! { fn wrapping_floor(self) -> Self }
             trait_delegate! { fn wrapping_round(self) -> Self }
             trait_delegate! { fn wrapping_round_ties_to_even(self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_ceil(self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_floor(self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_round(self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_round_ties_to_even(self) -> Self }
             trait_delegate! { fn overflowing_ceil(self) -> (Self, bool) }
             trait_delegate! { fn overflowing_floor(self) -> (Self, bool) }
@@ -2547,33 +2413,19 @@ macro_rules! impl_fixed {
             trait_delegate! { fn wrapping_rem_euclid_int(self, rhs: Self::Bits) -> Self }
             trait_delegate! { fn wrapping_shl(self, rhs: u32) -> Self }
             trait_delegate! { fn wrapping_shr(self, rhs: u32) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_neg(self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_add(self, rhs: Self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_sub(self, rhs: Self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_mul(self, rhs: Self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_div(self, rhs: Self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_recip(self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_mul_add(self, mul: Self, add: Self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_div_euclid(self, rhs: Self) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_mul_int(self, rhs: Self::Bits) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_div_int(self, rhs: Self::Bits) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_div_euclid_int(self, rhs: Self::Bits) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_rem_euclid_int(self, rhs: Self::Bits) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_shl(self, rhs: u32) -> Self }
-            #[cfg(feature = "unwrapped")]
             trait_delegate! { fn unwrapped_shr(self, rhs: u32) -> Self }
             trait_delegate! { fn overflowing_neg(self) -> (Self, bool) }
             trait_delegate! { fn overflowing_add(self, rhs: Self) -> (Self, bool) }
@@ -2754,9 +2606,7 @@ macro_rules! impl_fixed {
                 trait_delegate! { fn saturating_signum(self) -> Self }
                 trait_delegate! { fn wrapping_abs(self) -> Self }
                 trait_delegate! { fn wrapping_signum(self) -> Self }
-                #[cfg(feature = "unwrapped")]
                 trait_delegate! { fn unwrapped_abs(self) -> Self }
-                #[cfg(feature = "unwrapped")]
                 trait_delegate! { fn unwrapped_signum(self) -> Self }
                 trait_delegate! { fn overflowing_abs(self) -> (Self, bool) }
                 trait_delegate! { fn overflowing_signum(self) -> (Self, bool) }
@@ -2772,7 +2622,6 @@ macro_rules! impl_fixed {
                 trait_delegate! { fn next_power_of_two(self) -> Self }
                 trait_delegate! { fn checked_next_power_of_two(self) -> Option<Self> }
                 trait_delegate! { fn wrapping_next_power_of_two(self) -> Self }
-                #[cfg(feature = "unwrapped")]
                 trait_delegate! { fn unwrapped_next_power_of_two(self) -> Self }
             }
         }